Fix type

+++ Task3/Program/main.py
@@ -49,7 +49,7 @@ def main() -> None:
 
 # DEF ------------------------------------------------------------------------ #
 def process_classifiers(data_sets: Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray],
-                        data_set_name: str, save_latex: bool, test_data_size: int) -> None:
+                        data_set_name: str, save_latex: bool, test_data_size: float) -> None:
     display_header("k-nearest neighbors classifier")
     knn_classification(data_sets, True, data_set_name, save_latex)
     knn_classification(data_sets, False, data_set_name, save_latex)
